refactor(ui): extract ModalBase for dialog scaffolding
InfoDialog and QuizDialog each reimplemented the same overlay + fade +
Esc-key scaffolding. Lift that into an abstract ModalBase that owns
the container, the interactive overlay, fade-in/fade-out lifecycle,
and the Esc key binding. Subclasses now only build their panel content
and override onBeforeClose / onAfterClose.

Drops ~60 lines of duplicated scaffolding across the two dialogs.
No behavior change.
