+++ src/ui/ModalBase.ts
@@ -1,0 +1,89 @@
+import * as Phaser from 'phaser';
+import { GAME_WIDTH, GAME_HEIGHT } from '../config/gameConfig';
+
+/**
+ * Shared scaffolding for full-screen modal overlays (info + quiz dialogs).
+ *
+ * Owns the container, the dimmed overlay, the Esc-to-close keybinding,
+ * and the fade in/out lifecycle. Subclasses populate the container with
+ * their panel content in the constructor and call `fadeIn()` when ready.
+ *
+ * Depth 200 and scrollFactor 0 are standard for all in-game overlays.
+ */
+export abstract class ModalBase {
+  protected readonly scene: Phaser.Scene;
+  protected readonly container: Phaser.GameObjects.Container;
+
+  private escKey: Phaser.Input.Keyboard.Key | null = null;
+  private escHandler: (() => void) | null = null;
+  private destroyed = false;
+
+  constructor(scene: Phaser.Scene) {
+    this.scene = scene;
+
+    this.container = scene.add.container(0, 0);
+    this.container.setDepth(200);
+    this.container.setScrollFactor(0);
+    this.container.setAlpha(0);
+
+    this.buildOverlay();
+    this.registerEscKey();
+  }
+
+  /** Dimmed fullscreen rect; added as the first child so subclasses can rely on index 0. */
+  private buildOverlay(): void {
+    const overlay = this.scene.add.rectangle(
+      GAME_WIDTH / 2, GAME_HEIGHT / 2,
+      GAME_WIDTH, GAME_HEIGHT,
+      0x000000, 0.65,
+    );
+    overlay.setScrollFactor(0).setInteractive(); // block clicks through
+    this.container.add(overlay);
+  }
+
+  private registerEscKey(): void {
+    if (!this.scene.input.keyboard) return;
+    this.escKey = this.scene.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.ESC);
+    this.escHandler = () => {
+      if (this.escKey?.isDown) this.close();
+    };
+    this.scene.events.on('update', this.escHandler);
+  }
+
+  /** Call at end of subclass constructor once panel content is built. */
+  protected fadeIn(duration = 200): void {
+    this.scene.tweens.add({ targets: this.container, alpha: 1, duration });
+  }
+
+  /** Hook for subclasses to release additional resources before the fade-out. */
+  protected onBeforeClose(): void {
+    /* default no-op */
+  }
+
+  /** Hook fired after the container is destroyed. Subclasses forward to their onClose callback. */
+  protected onAfterClose(): void {
+    /* default no-op */
+  }
+
+  close(): void {
+    if (this.destroyed) return;
+    this.destroyed = true;
+
+    this.onBeforeClose();
+
+    if (this.escHandler) {
+      this.scene.events.off('update', this.escHandler);
+    }
+    if (this.escKey) {
+      this.escKey.destroy();
+    }
+
+    this.scene.tweens.add({
+      targets: this.container, alpha: 0, duration: 150,
+      onComplete: () => {
+        this.container.destroy();
+        this.onAfterClose();
+      },
+    });
+  }
+}
